{"schema":"decodex/commit/1","summary":"Move window redraw requests","authority":"manual"}
